% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morrr-package.R
\docType{package}
\name{morrr-package}
\alias{morrr}
\alias{morrr-package}
\title{morrr: Page Console Output Like Unix More}
\description{
Page console output in the R Console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adam-gruer/morrr}
  \item Report bugs at \url{https://github.com/adam-gruer/morrr/issues}
}

}
\author{
\strong{Maintainer}: Adam Gruer \email{adamgruer@icloud.com}

}
\keyword{internal}
